// src/services/central-api/generated/model/ctl-mst-fee-rate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * PL Cental API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface CtlMstFeeRate
 */
export interface CtlMstFeeRate {
    /**
     * 
     * @type {string}
     * @memberof CtlMstFeeRate
     */
    'code'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlMstFeeRate
     */
    'name'?: string | null;
    /**
     * 
     * @type {number}
     * @memberof CtlMstFeeRate
     */
    'amount'?: number | null;
    /**
     * 
     * @type {string}
     * @memberof CtlMstFeeRate
     */
    'createTimestamp'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlMstFeeRate
     */
    'createBy'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlMstFeeRate
     */
    'updateTimestamp'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlMstFeeRate
     */
    'updateBy'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlMstFeeRate
     */
    'deleteTimestamp'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlMstFeeRate
     */
    'deleteBy'?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof CtlMstFeeRate
     */
    'isDelete'?: boolean | null;
    /**
     * 
     * @type {string}
     * @memberof CtlMstFeeRate
     */
    'tenantCode'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlMstFeeRate
     */
    'branchCode'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CtlMstFeeRate
     */
    'status'?: string | null;
}